requirements.txt
torch
torchvision
pillow
tqdm
diffusers
transformers
einops 